# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenant/members:
    get:
      operationId: listTenantMembers
      tags: [membership]
      summary: List members of the active customer
      responses:
        "200":
          description: Customer members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantMember" }
        "403": { description: Customer administrator permission required }
  /tenant/members/{userId}:
    patch:
      operationId: updateTenantMember
      tags: [membership]
      summary: Change a customer's member role
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTenantMemberInput" }
      responses:
        "200":
          description: Updated membership
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantMember" }
        "403": { description: Customer administrator permission required }
        "409": { description: The last owner cannot be removed }
    delete:
      operationId: removeTenantMember
      tags: [membership]
      summary: Remove a member from the active customer
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: Member removed }
        "403": { description: Customer administrator permission required }
        "409": { description: The last owner cannot be removed }
  /tenant/invitations:
    get:
      operationId: listTenantInvitations
      tags: [membership]
      summary: List invitations for the active customer
      responses:
        "200":
          description: Customer invitations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TenantInvitation" }
        "403": { description: Customer administrator permission required }
    post:
      operationId: createTenantInvitation
      tags: [membership]
      summary: Invite a user to the active customer
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTenantInvitationInput" }
      responses:
        "201":
          description: Invitation created; the token is returned only once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedTenantInvitation" }
        "403": { description: Customer administrator permission required }
        "409": { description: Active invitation already exists }
  /tenant/invitations/{invitationId}/revoke:
    post:
      operationId: revokeTenantInvitation
      tags: [membership]
      summary: Revoke a pending customer invitation
      parameters:
        - in: path
          name: invitationId
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Revoked invitation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TenantInvitation" }
        "403": { description: Customer administrator permission required }
...
components:
...
  schemas:
...
    TenantMember:
      type: object
      required: [userId, email, displayName, role, joinedAt]
      properties:
        userId: { type: integer }
        email: { type: ["string", "null"] }
        displayName: { type: ["string", "null"] }
        role: { type: string, enum: [owner, admin, member, viewer] }
        joinedAt: { type: string, format: date-time }
    UpdateTenantMemberInput:
      type: object
      required: [role]
      properties:
        role: { type: string, enum: [owner, admin, member, viewer] }
    CreateTenantInvitationInput:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [owner, admin, member, viewer] }
    TenantInvitation:
      type: object
      required: [id, tenantId, email, role, status, expiresAt, createdAt]
      properties:
        id: { type: integer }
        tenantId: { type: integer }
        email: { type: string, format: email }
        role: { type: string, enum: [owner, admin, member, viewer] }
        status: { type: string, enum: [pending, accepted, revoked, expired] }
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    CreatedTenantInvitation:
      type: object
      required: [invitation, token]
      properties:
        invitation: { $ref: "#/components/schemas/TenantInvitation" }
        token: { type: string }
...