# lib/api-zod/src/generated/types/tenantMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TenantMemberRole } from './tenantMemberRole';

export interface TenantMember {
  userId: number;
  /** @nullable */
  email: string | null;
  /** @nullable */
  displayName: string | null;
  role: TenantMemberRole;
  joinedAt: Date;
}